fstools: block: Add lvm volume status change scan on hotplug
This change adds a call to pvscan --cache -aay on hot-plug in of LVM
volumes and pvscan --cache on plug-out.

That should be sufficient to trigger the dm-xx events required to
mount the new volume.

NOTE:  This change does not modify system startup or shutdown operation
e.g. block mount or block unmount so the lvm2 service start and stop
is still required.  To work without the service the scan would
have to be triggered at the mount/umount operations rather than
mount_action.  A delay and cache reload would likely also be required
so that does not seem ideal.

Signed-off-by: David Adair <djabhead@aol.com>