fstools: block: Don't umount internal overlay.
block umount avoids / and /overlay unless -a is specified.

This change adds /rom and /rom/overlay to the list of
avoided paths so internal and external overlay work
the same.

NOTE: overlayfs does not seem to care if it is layered
on top of un-mounted fs -- it just hangs out in
BLK_DETACH state forever.  This is a pre-emptive change
to avoid issues if:
 - We try to clean up the detached FS entries e.g. to
free up md maps.  It is simpler if detached==needs cleanup.
 - Someone decides to use remount,ro, umount incantation
on theory root is "safe" due to existing logic.

Besides symetry is pretty....

Signed-off-by: David Adair <djabhead@aol.com>