fstools: blockd: Support hotplug-remove for LVM volumes.
Block and therefore hotplug prefers /dev/mapper entries
to /dev/dm-xxx entries.  The result is that our device
list will be based on the /dev/mapper names.

Make sure we examine that path as well as /dev when
searching for a matching device during the hotplug
remove callbacks or the volumes will not be removed.

NOTE:  This will still not work if the requested
device is "dm-0" -- it only really works for autofs
devices with volume name == device.  That is sensible
because UBUS should only be used for items we report
in block info.  For other hotplug requests they should
go to block first and it should sort out the mapping.

Signed-off-by: David Adair <djabhead@aol.com>